+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ Heat Pump Display Release Notes
 - Add statistics screen to display some values of the DTA logging file of the controller.
 - Add heat quantities.
 - Add compressor frequency.
+- Add info about software updates, the new web interface and the repo of this app.
 
 ## 16.0.1 (2022-03-11)